Added ViewModel, LiveData Observer, Journal alert dialog
Journal data is now apart of the viewmodel using livedata and observer so exercises wont be deleted after fragment is destroyed (navigating fragments). This will also allow for easier data implementation in the future. Pressing on a list item in Journal now opens up a dialog popup for recording sets reps and weight.